Rename SlackBotPingWorker to Slack::Messengers::Worker
