+++ app/services/pro_memberships/expiration_notifier.rb
@@ -27,7 +27,7 @@ module ProMemberships
         membership.increment(:expiration_notifications_count)
         membership.save!
 
-        SlackBotPingWorker.perform_async(
+        Slack::Messengers::Worker.perform_async(
           message: "#{membership.user.name}'s pro membership expires on #{expiration_date}",
           channel: "pro-memberships",
           username: "pro-memberships",
