+++ app/services/slack/messengers/feedback.rb
@@ -42,7 +42,7 @@ module Slack
           message: message,
         )
 
-        SlackBotPingWorker.perform_async(
+        Slack::Messengers::Worker.perform_async(
           message: final_message,
           channel: type,
           username: "#{type}_bot",
