+++ spec/models/comment_spec.rb
@@ -324,19 +324,19 @@ RSpec.describe Comment, type: :model do
 
       before do
         # making sure there are no other enqueued jobs from other tests
-        sidekiq_perform_enqueued_jobs(only: SlackBotPingWorker)
+        sidekiq_perform_enqueued_jobs(only: Slack::Messengers::Worker)
       end
 
       it "queues a slack message when a warned user leaves a comment" do
         user.add_role(:warned)
 
-        sidekiq_assert_enqueued_jobs(1, only: SlackBotPingWorker) do
+        sidekiq_assert_enqueued_jobs(1, only: Slack::Messengers::Worker) do
           create(:comment, user: user, commentable: article)
         end
       end
 
       it "does not send notification if a regular user leaves a comment" do
-        sidekiq_assert_no_enqueued_jobs(only: SlackBotPingWorker) do
+        sidekiq_assert_no_enqueued_jobs(only: Slack::Messengers::Worker) do
           create(:comment, commentable: article, user: user)
         end
       end
