+++ spec/models/reaction_spec.rb
@@ -187,23 +187,23 @@ RSpec.describe Reaction, type: :model do
 
       before do
         # making sure there are no other enqueued jobs from other tests
-        sidekiq_perform_enqueued_jobs(only: SlackBotPingWorker)
+        sidekiq_perform_enqueued_jobs(only: Slack::Messengers::Worker)
       end
 
       it "queues a slack message to be sent for a vomit reaction" do
-        sidekiq_assert_enqueued_jobs(1, only: SlackBotPingWorker) do
+        sidekiq_assert_enqueued_jobs(1, only: Slack::Messengers::Worker) do
           create(:reaction, reactable: article, user: user, category: "vomit")
         end
       end
 
       it "does not queue a message for a like reaction" do
-        sidekiq_assert_no_enqueued_jobs(only: SlackBotPingWorker) do
+        sidekiq_assert_no_enqueued_jobs(only: Slack::Messengers::Worker) do
           create(:reaction, reactable: article, user: user, category: "like")
         end
       end
 
       it "does not queue a message for a thumbsdown reaction" do
-        sidekiq_assert_no_enqueued_jobs(only: SlackBotPingWorker) do
+        sidekiq_assert_no_enqueued_jobs(only: Slack::Messengers::Worker) do
           create(:reaction, reactable: article, user: user, category: "thumbsdown")
         end
       end
