+++ spec/services/slack/messengers/article_fetched_feed_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe Slack::Messengers::ArticleFetchedFeed, type: :service do
   let(:default_params) { { article: article } }
 
   it "does not message slack for a new article not coming from the feed" do
-    sidekiq_assert_no_enqueued_jobs(only: SlackBotPingWorker) do
+    sidekiq_assert_no_enqueued_jobs(only: Slack::Messengers::Worker) do
       article.published_from_feed = false
 
       described_class.call(article: article)
@@ -19,11 +19,11 @@ RSpec.describe Slack::Messengers::ArticleFetchedFeed, type: :service do
   end
 
   it "contains the correct info", :aggregate_failures do
-    sidekiq_assert_enqueued_jobs(1, only: SlackBotPingWorker) do
+    sidekiq_assert_enqueued_jobs(1, only: Slack::Messengers::Worker) do
       described_class.call(default_params)
     end
 
-    job = sidekiq_enqueued_jobs(worker: SlackBotPingWorker).last
+    job = sidekiq_enqueued_jobs(worker: Slack::Messengers::Worker).last
     message = job["args"].first["message"]
 
     expect(message).to include(article.title)
@@ -31,11 +31,11 @@ RSpec.describe Slack::Messengers::ArticleFetchedFeed, type: :service do
   end
 
   it "messages the proper channel with the proper username and emoji", :aggregate_failures do
-    sidekiq_assert_enqueued_jobs(1, only: SlackBotPingWorker) do
+    sidekiq_assert_enqueued_jobs(1, only: Slack::Messengers::Worker) do
       described_class.call(default_params)
     end
 
-    job = sidekiq_enqueued_jobs(worker: SlackBotPingWorker).last
+    job = sidekiq_enqueued_jobs(worker: Slack::Messengers::Worker).last
     job_args = job["args"].first
 
     expect(job_args["channel"]).to eq("activity")
