+++ spec/services/slack/messengers/article_published_spec.rb
@@ -12,14 +12,14 @@ RSpec.describe Slack::Messengers::ArticlePublished, type: :service do
   let(:default_params) { { article: article } }
 
   it "does not message slack for a draft article" do
-    sidekiq_assert_no_enqueued_jobs(only: SlackBotPingWorker) do
+    sidekiq_assert_no_enqueued_jobs(only: Slack::Messengers::Worker) do
       article = build(:article, published: false)
       described_class.call(article: article)
     end
   end
 
   it "does not message slack for an article that was published long ago" do
-    sidekiq_assert_no_enqueued_jobs(only: SlackBotPingWorker) do
+    sidekiq_assert_no_enqueued_jobs(only: Slack::Messengers::Worker) do
       article = build(:article).tap do |art|
         art.published = true
         art.published_at = 1.minute.ago
@@ -29,11 +29,11 @@ RSpec.describe Slack::Messengers::ArticlePublished, type: :service do
   end
 
   it "contains the correct info", :aggregate_failures do
-    sidekiq_assert_enqueued_jobs(1, only: SlackBotPingWorker) do
+    sidekiq_assert_enqueued_jobs(1, only: Slack::Messengers::Worker) do
       described_class.call(default_params)
     end
 
-    job = sidekiq_enqueued_jobs(worker: SlackBotPingWorker).last
+    job = sidekiq_enqueued_jobs(worker: Slack::Messengers::Worker).last
     message = job["args"].first["message"]
 
     expect(message).to include(article.title)
@@ -41,11 +41,11 @@ RSpec.describe Slack::Messengers::ArticlePublished, type: :service do
   end
 
   it "messages the proper channel with the proper username and emoji", :aggregate_failures do
-    sidekiq_assert_enqueued_jobs(1, only: SlackBotPingWorker) do
+    sidekiq_assert_enqueued_jobs(1, only: Slack::Messengers::Worker) do
       described_class.call(default_params)
     end
 
-    job = sidekiq_enqueued_jobs(worker: SlackBotPingWorker).last
+    job = sidekiq_enqueued_jobs(worker: Slack::Messengers::Worker).last
     job_args = job["args"].first
 
     expect(job_args["channel"]).to eq("activity")
