+++ spec/services/slack/messengers/comment_user_warned_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Slack::Messengers::CommentUserWarned, type: :service do
   let(:default_params) { { comment: comment } }
 
   it "does not message slack for a comment with a regular user" do
-    sidekiq_assert_no_enqueued_jobs(only: SlackBotPingWorker) do
+    sidekiq_assert_no_enqueued_jobs(only: Slack::Messengers::Worker) do
       described_class.call(comment: build(:comment, user: build(:user)))
     end
   end
@@ -18,11 +18,11 @@ RSpec.describe Slack::Messengers::CommentUserWarned, type: :service do
     end
 
     it "contains the correct info", :aggregate_failures do
-      sidekiq_assert_enqueued_jobs(1, only: SlackBotPingWorker) do
+      sidekiq_assert_enqueued_jobs(1, only: Slack::Messengers::Worker) do
         described_class.call(default_params)
       end
 
-      job = sidekiq_enqueued_jobs(worker: SlackBotPingWorker).last
+      job = sidekiq_enqueued_jobs(worker: Slack::Messengers::Worker).last
       message = job["args"].first["message"]
 
       internal_user_url = URL.url(
@@ -36,11 +36,11 @@ RSpec.describe Slack::Messengers::CommentUserWarned, type: :service do
     end
 
     it "messages the proper channel with the proper username and emoji", :aggregate_failures do
-      sidekiq_assert_enqueued_jobs(1, only: SlackBotPingWorker) do
+      sidekiq_assert_enqueued_jobs(1, only: Slack::Messengers::Worker) do
         described_class.call(default_params)
       end
 
-      job = sidekiq_enqueued_jobs(worker: SlackBotPingWorker).last
+      job = sidekiq_enqueued_jobs(worker: Slack::Messengers::Worker).last
       job_args = job["args"].first
 
       expect(job_args["channel"]).to eq("warned-user-comments")
