+++ spec/services/slack/messengers/reaction_vomit_spec.rb
@@ -7,18 +7,18 @@ RSpec.describe Slack::Messengers::ReactionVomit, type: :service do
   let(:default_params) { { reaction: reaction } }
 
   it "does not message slack for a like reaction" do
-    sidekiq_assert_no_enqueued_jobs(only: SlackBotPingWorker) do
+    sidekiq_assert_no_enqueued_jobs(only: Slack::Messengers::Worker) do
       reaction = build(:reaction, category: :like)
       described_class.call(reaction: reaction)
     end
   end
 
   it "contains the correct info", :aggregate_failures do
-    sidekiq_assert_enqueued_jobs(1, only: SlackBotPingWorker) do
+    sidekiq_assert_enqueued_jobs(1, only: Slack::Messengers::Worker) do
       described_class.call(default_params)
     end
 
-    job = sidekiq_enqueued_jobs(worker: SlackBotPingWorker).last
+    job = sidekiq_enqueued_jobs(worker: Slack::Messengers::Worker).last
     message = job["args"].first["message"]
 
     expect(message).to include(user.name)
@@ -28,11 +28,11 @@ RSpec.describe Slack::Messengers::ReactionVomit, type: :service do
   end
 
   it "messages the proper channel with the proper username and emoji", :aggregate_failures do
-    sidekiq_assert_enqueued_jobs(1, only: SlackBotPingWorker) do
+    sidekiq_assert_enqueued_jobs(1, only: Slack::Messengers::Worker) do
       described_class.call(default_params)
     end
 
-    job = sidekiq_enqueued_jobs(worker: SlackBotPingWorker).last
+    job = sidekiq_enqueued_jobs(worker: Slack::Messengers::Worker).last
     job_args = job["args"].first
 
     expect(job_args["channel"]).to eq("abuse-reports")
